// src/app/models/RegisterModel.ts
export interface RegisterModel {
  nombre: string;
  paterno: string;
  materno?: string;
  fechaNacimiento: string;
  telefono: string;
  email: string;
  password: string;
  nombreNegocio: string;
  direccion: string;
  telefonoNegocio: string;
  tipoNegocio: string;
  plan: string;
  status: string;
  token: string;
}